Renaming `index.tsx` files to their real component name
* And `styles.tsx` to `[compnentName].styles.tsx`
* And `service.tsx` to `[compnentName].service.tsx`

+ Schemas/Movie.js -> Schemas/MovieComment.js
