+++ client/components/App.tsx
@@ -9,16 +9,16 @@ import enTranslation from "../translations/en.json";
 import frTranslation from "../translations/fr.json";
 
 import CustomRoute from "./Routes/CustomRoute";
-import Layout from "./Layout";
+import Layout from "./Layout/Layout";
 import ConfirmEmail from "./Authentication/ConfirmEmail";
-import Home from "./Home";
+import Home from "./Home/Home";
 import Movie from "./Movie/Movie";
-import Search from "./Search";
-import FourOhFour from "./FourOhFour";
+import Search from "./Search/Search";
+import FourOhFour from "./FourOhFour/FourOhFour";
 import SignIn from "./Authentication/SignIn";
 import SignUp from "./Authentication/SignUp";
 import ResetPassword from "./Authentication/ResetPassword";
-import Error from "./Error";
+import Error from "./Error/Error";
 
 import useLocaleStorage from "../hooks/useLocaleStorage";
