+++ client/components/Layout/FiltersSelect.tsx
@@ -3,7 +3,7 @@ import { useIntl } from "react-intl";
 
 import { MenuItem, Select } from "@material-ui/core";
 
-import { useFiltersStyles } from "./styles";
+import { useFiltersStyles } from "./Layout.styles";
 
 interface Props {
   collections: string;
